b00463f QtAutomoc test: Pass QT_QMAKE_EXECUTABLE
e78ce44 Fix automoc with VS builds: apply patch from Bill
71165e9 Silence warning in automoc: use long instead of int

+++ Source/cmQtAutomoc.cxx
@@ -519,7 +519,7 @@ void cmQtAutomoc::ParseCppFile(const std::string& absFilename,
   const std::string absPath = cmsys::SystemTools::GetFilenamePath(
                    cmsys::SystemTools::GetRealPath(absFilename.c_str())) + '/';
 
-  int matchOffset = 0;
+  long matchOffset = 0;
   if (!mocIncludeRegExp.find(contentsString.c_str()))
     {
     // no moc #include, look whether we need to create a moc from
